kmsv2: unit tests same package as implementation
